auth_service/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
gunicorn==22.0.0
PyJWT==2.8.0
